requirements.txt
Django==2.1.7
pytz==2018.9
gunicorn==19.9.0
django-heroku==0.3.1
dj-database-url==0.5.0
python-decouple==3.1
